+++ src/components/DashboardPeriodSelector.vue
@@ -1,24 +1,21 @@
 <template>
   <div v-click-outside="hide" class="periodSelector">
-    <div
-      class="periodSelector__input"
-      :class="{ 'periodSelector__input--active': isOpen }"
-      tabindex="0"
+    <button
+      class="pkt-btn pkt-btn--flat pkt-btn--small pkt-btn--icon-right"
+      :class="{ 'pkt-btn--focus': isOpen }"
       @click="toggle"
-      @keyup.enter="toggle"
     >
-      <span class="periodSelector__input-value">
-        {{ rangeLabel }}
-      </span>
-      <pkt-icon name="calendar" />
-    </div>
+      <pkt-icon name="calendar" class="pkt-btn__icon" />
+      <span class="pkt-btn__text">{{ label }}</span>
+    </button>
     <div v-if="isOpen" class="periodSelector__content">
       <button
         v-for="rangeOption in options"
         :key="rangeOption.value"
         class="periodSelector__option"
         :class="{
-          'periodSelector__option--active': rangeOption.key === period.key,
+          'periodSelector__option--active':
+            selectedPeriod && rangeOption.key === selectedPeriod.key,
         }"
         @click="selectRangeOption(rangeOption)"
       >
@@ -37,39 +34,18 @@
 </template>
 
 <script>
+import { mapState, mapActions } from 'vuex';
 import ClickOutside from 'vue-click-outside';
 import locale from 'flatpickr/dist/l10n/no';
 import endOfDay from 'date-fns/endOfDay';
 import { dateLongCompact } from '@/util';
+import getPeriods from '@/config/periods';
 
 export default {
   name: 'DashboardPeriodSelector',
 
   directives: {
     ClickOutside,
-  },
-
-  props: {
-    period: {
-      type: Object,
-      required: false,
-      default: null,
-    },
-    options: {
-      type: Array,
-      required: false,
-      default: null,
-    },
-    startDate: {
-      type: Date,
-      required: false,
-      default: null,
-    },
-    endDate: {
-      type: Date,
-      required: false,
-      default: null,
-    },
   },
 
   data: () => ({
@@ -82,43 +58,41 @@ export default {
       locale: locale.no,
     },
     range: null,
-    formattedRangeLabel: null,
+    options: Object.values(getPeriods()),
   }),
 
   computed: {
-    rangeLabel() {
-      const selectedPeriodOption = this.options.find(
-        (option) => option.key === this.period.key
-      );
-      return selectedPeriodOption?.label || this.formattedRangeLabel;
+    ...mapState(['selectedPeriod']),
+
+    label() {
+      if (!this.selectedPeriod) {
+        return this.$t('period.choosePeriod');
+      }
+      if (!this.selectedPeriod.label) {
+        if (Array.isArray(this.range) && this.range.filter((d) => d).length === 2) {
+          return [...new Set(this.range.map(dateLongCompact))].join(
+            this.flatPickerConfig.locale.rangeSeparator
+          );
+        }
+      }
+      return this.selectedPeriod.label;
     },
   },
 
   watch: {
-    startDate: {
-      immediate: true,
-      async handler() {
-        this.range = [this.startDate, this.endDate];
-      },
-    },
-
-    endDate: {
-      immediate: true,
-      async handler() {
-        this.range = [this.startDate, this.endDate];
-      },
-    },
-
-    range(range) {
-      if (Array.isArray(range) && range.filter((d) => d).length === 2) {
-        this.formattedRangeLabel = [...new Set(range.map(dateLongCompact))].join(
-          this.flatPickerConfig.locale.rangeSeparator
-        );
-      }
+    isOpen() {
+      // Reset the internal range property to the currenct period selection when
+      // toggling the picker. This to ensure that no invalid (incomplete) range
+      // selections are kept in the component state, and that the range is set
+      // correctly when navigation back to the parent view.
+      const { startDate, endDate } = this.selectedPeriod;
+      this.range = startDate && endDate ? [startDate, endDate] : null;
     },
   },
 
   methods: {
+    ...mapActions(['setSelectedPeriod']),
+
     toggle() {
       this.isOpen = !this.isOpen;
     },
@@ -127,17 +101,21 @@ export default {
     },
     selectRangeOption(rangeOption) {
       this.range = [rangeOption.startDate, rangeOption.endDate];
-      this.$emit('input', rangeOption);
+      this.setSelectedPeriod(rangeOption);
       this.hide();
     },
     selectCustomRange(range) {
       if (range.length !== 2) {
         return;
       }
-      this.$emit('input', {
+      this.setSelectedPeriod({
         startDate: range[0],
         endDate: endOfDay(range[1]),
+        label: [...new Set(range.map(dateLongCompact))].join(
+          this.flatPickerConfig.locale.rangeSeparator
+        ),
       });
+      this.hide();
     },
   },
 };
@@ -147,35 +125,12 @@ export default {
 .periodSelector {
   position: relative;
   display: inline-block;
-}
 
-.periodSelector__input {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-  padding: 0.75rem;
-  white-space: nowrap;
-  border: 1px solid var(--color-grayscale-10);
-  cursor: pointer;
-
-  > svg {
-    height: 1.25rem;
-  }
-
-  &:hover:not(&--active) {
-    background: var(--color-gray-light);
-    border-color: var(--color-gray-light);
-  }
-
-  &--active {
-    background-color: var(--color-secondary-light);
-    border-color: var(--color-secondary-light);
-  }
-
-  &-value {
-    color: var(--color-text);
-    font-weight: 500;
-    font-size: 1rem;
+  .pkt-btn--focus {
+    // Somewhat alter the focus state to more clearly indicate
+    // that the button is in focused state while dropdown is open.
+    text-decoration: none;
+    border-color: var(--btn-focus-txt);
   }
 }